requirements.txt
django
pillow
django-taggit
django-shortuuidfield
django-shopping-cart